// src/generated/model/createClassRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { CreateClassRequestStatus } from './createClassRequestStatus.ts';

export interface CreateClassRequest {
  programId: string;
  programLevelId: string;
  batchId: string;
  /**
     * @minLength 0
     * @maxLength 100
     */
  name: string;
  /** @minimum 1 */
  capacity: number;
  status: CreateClassRequestStatus;
  remarks?: string;
}