// src/generated/model/updateClassRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { UpdateClassRequestStatus } from './updateClassRequestStatus.ts';

export interface UpdateClassRequest {
  /**
     * @minLength 0
     * @maxLength 100
     */
  name: string;
  /** @minimum 1 */
  capacity: number;
  status: UpdateClassRequestStatus;
  remarks?: string;
}